fixed inaccurate EV on shoe games added SD H17 strategy some UI improvements removed S17 option since the strategies are not implemented
