+++ BJ/PlayingStrategies/ShoeS17.cs
@@ -1,0 +1,7 @@
+﻿namespace BJ.PlayingStrategies
+{
+	public static partial class PlayingStrategies
+	{
+
+	}
+}
