+++ BJ/Program.cs
@@ -1,11 +1,12 @@
 ﻿using BJ;
+using BJ.PlayingStrategies;
 
 const decimal INITIALBANKROLL = 2000;
 const int ROUNDS = 1000;
 const int ROUNDSPERHOUR = 100;
 const int CONCURRENTPLAYERS = 10000;
 const int BETTINGUNIT = 1;
-const decimal DECKPEN = 0.70m;
+const int DECKPEN = 70;
 const int SHOESIZE = 2;
 const decimal BJPAYOUT = 1.5m;
 const int MAXHANDS = 4;
@@ -34,7 +35,7 @@ Parallel.For(0, CONCURRENTPLAYERS, i =>
 		INITIALBANKROLL,
 		//new BettingStrategy(BettingStrategies.BuildBetSpread([0, 0, 25], [25, 25, 50, 75, 125, 150, 200]), [1, 1, 2]),
 		new BettingStrategy(BettingStrategies.BuildBetSpread([0, 0, 10], [10, 10, 20, 30, 40, 50]), [1, 2]),
-		new PlayingStrategy_2D().UseDeviations()
+		new PlayingStrategy_DD_H17().UseDeviations()
 	);
 
 	var bjplayer = new BJPlayer(table, player);
@@ -76,7 +77,7 @@ static decimal PlayBlackjack(BJPlayer bjplayer, int bettingunit, int rounds)
 
 		var (bet, hands) = player.GetBetAmount((int)bjplayer.Table.GetTrueCount());
 #if DEBUG
-        Console.WriteLine($"Betting ${bet} on {hands} hands.");
+		Console.WriteLine($"Betting ${bet} on {hands} hands.");
 #endif
 
 		player.AddHands(hands);
@@ -87,14 +88,14 @@ static decimal PlayBlackjack(BJPlayer bjplayer, int bettingunit, int rounds)
 		if (bj.DoReshuffleShoe())
 		{
 #if DEBUG
-            Console.WriteLine($"Refilling and shuffling shoe!");
+			Console.WriteLine($"Refilling and shuffling shoe!");
 #endif
 			bj.RefillShoe();
 		}
 
 #if DEBUG
-        Console.WriteLine($"Player's new Bankroll is ${player.BankRoll}.");
-        Console.WriteLine($"RC is {bj.RunningCount} and TC is {bj.GetTrueCount()}.\n\n");
+		Console.WriteLine($"Player's new Bankroll is ${player.BankRoll}.");
+		Console.WriteLine($"RC is {bj.RunningCount} and TC is {bj.GetTrueCount()}.\n\n");
 #endif
 	}
